docker/localizer/pyproject.toml
[project]
name = "plerion-localizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "common",
  "pycolmap>=3.13.0",
  "h5py>=3.15.0",
  "fastapi>=0.119.0",
  "pydantic-settings>=2.11.0",
  "python-multipart>=0.0.20",
  "uvicorn[standard]>=0.35.0",
  "opencv-python>=4.10.0",
  "scipy>=1.16.2",
  "faiss-cpu>=1.12.0",
  "core[colmap]",
]

[dependency-groups]
dev = ["h5py-stubs>=0.1.2", "scipy-stubs>=1.16.2.4"]
torch = [
  "torch==2.6.0+cu126",
  "torchvision==0.21.0+cu126",
  "scikit-learn>=1.7.2",
  # https://github.com/astral-sh/uv/issues/14237
  # https://discuss.pytorch.org/t/when-i-install-torch-2-6-0-with-whl-cu126-none-of-cuda-dependencies-get-installed-i-have-cu126-in-the-environment-already-is-this-expected/218014/3
  "nvidia-cuda-nvrtc-cu12==12.6.77; platform_system == 'Linux'",
  "nvidia-cuda-runtime-cu12==12.6.77; platform_system == 'Linux'",
  "nvidia-cuda-cupti-cu12==12.6.80; platform_system == 'Linux'",
  "nvidia-cudnn-cu12==9.5.1.17; platform_system == 'Linux'",
  "nvidia-cublas-cu12==12.6.4.1; platform_system == 'Linux'",
  "nvidia-cufft-cu12==11.3.0.4; platform_system == 'Linux'",
  "nvidia-curand-cu12==10.3.7.77; platform_system == 'Linux'",
  "nvidia-cusolver-cu12==11.7.1.2; platform_system == 'Linux'",
  "nvidia-cusparse-cu12==12.5.4.2; platform_system == 'Linux'",
  "nvidia-cusparselt-cu12==0.6.3; platform_system == 'Linux'",
  "nvidia-nvtx-cu12==12.6.77; platform_system == 'Linux'",
  "nvidia-nvjitlink-cu12==12.6.85; platform_system == 'Linux'",
  "nvidia-nccl-cu12==2.21.5; platform_system == 'Linux'",
]
neural-networks = ["neural-networks"]

[tool.uv.sources]
core = { workspace = true }
common = { workspace = true }
neural-networks = { workspace = true }
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/localize"]

[project.scripts]
localize = "localize.main:main"
